refactor generate_file_list to helpers.py
