+++ client/py_client/ipfs.py
@@ -1,11 +1,10 @@
 import subprocess
-import warnings
 import os
 import requests
 import json
 
-from .helpers import take_only_last_cid
-from pathlib import Path
+from .helpers import take_only_last_cid, generate_file_list
+
 ICONS_PATH = './test-data/icons'
 use_ipfs_gateway = True
 try:
@@ -37,7 +36,7 @@ class Ipfs:
                 params = (
                     ('pin', 'true'),
                 )
-            files = Ipfs.generate_file_list(path_to_files)
+            files = generate_file_list(path_to_files)
             auth = ipfs_api_key.split(":")
             response = requests.post('https://ipfs.infura.io:5001/api/v0/add', headers=headers, params=params, files=files, auth=(auth[0], auth[1]))
 
@@ -57,21 +56,4 @@ class Ipfs:
             return ["QmP2fzfikh7VqTu8pvzd2G2vAd4eK7EaazXTEgqGN6AWoD"]
         return [Ipfs.add_recursive(f, local) for f in paths]
 
-
-    @staticmethod
-    def generate_file_list(path_to_files):
-        args = []
-        if os.path.isdir(path_to_files):
-            for dir_, _, files in os.walk(path_to_files):
-                for file_name in files:
-                    rel_path = os.path.relpath(os.path.join(dir_, file_name), str(Path(path_to_files).parent))
-                    rel_path = Path(rel_path)
-                    with open(os.path.join(dir_, file_name), 'rb') as file:
-                        args += [(rel_path.as_posix(), file.read())]
-        else:
-            rel_path = ''
-            rel_path = Path(rel_path)
-            with open(os.path.abspath(path_to_files), 'rb') as file:
-                args += [(rel_path.as_posix(), file.read())]
-        return args
